% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotCircle.R
\name{plotCircle}
\alias{plotCircle}
\title{Plot a circle of radius r}
\usage{
plotCircle(r = 1)
}
\arguments{
\item{r}{The radius of the circle to
   be plotted. Note the indentation on the
   extra description}
}
\value{
None The function is called solely
   for its side effect of drawing
}
\description{
This would be the more detailed description, if we
needed one.
}
\examples{
plotCircle(10)
}
\seealso{
The general
  \code{\link[graphics]{plot}} function
}

